Add logger to handle primary column exceptions
Refactored the code to add a logger to handle exceptions when identifying a primary column for a table. The changes include incorporating an instance of LoggerStrategy in the JunctionTable and CollaboratorsProgramsTable classes. LoggerStrategy instance is then used to log exceptions when the primary column couldn't be found in the primary table.